requirements.txt
Flask==1.1.2
numpy==1.18.5
torch==1.7.1
Flask_Cors==3.0.9
torchvision==0.8.2
Pillow==8.1.0
